+++ modules/post-processing/yolop-lane-detection/src/detect_lanes.py
@@ -1,13 +1,12 @@
+# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# SPDX-License-Identifier: Apache-2.0
+
 import argparse
-import csv
 import json
 import os
-import shutil
 import sys
 import time
 from pathlib import Path
-
-import imageio
 
 BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 sys.path.append(BASE_DIR)
@@ -15,15 +14,13 @@ sys.path.append(BASE_DIR)
 print(sys.path)
 import cv2
 import numpy as np
-import PIL.Image as image
-import scipy.special
+import pandas as pd
 import torch
 import torch.backends.cudnn as cudnn
 import torchvision.transforms as transforms
 from lib.config import cfg, update_config
 from lib.core.function import AverageMeter
 from lib.core.general import non_max_suppression, scale_coords
-from lib.core.postprocess import connect_lane, morphological_process
 from lib.dataset import LoadImages, LoadStreams
 from lib.models import get_net
 from lib.utils import plot_one_box, show_seg_result
@@ -41,14 +38,21 @@ transform = transforms.Compose(
 )
 
 
+class NumpyEncoder(json.JSONEncoder):
+    def default(self, obj):
+        if isinstance(obj, np.ndarray):
+            return obj.tolist()
+        return json.JSONEncoder.default(self, obj)
+
+
+name_lanes = []
+
+
 def detect(cfg, opt):
 
     logger, _, _ = create_logger(cfg, cfg.LOG_DIR, "demo")
 
     device = select_device(logger, opt.device)
-    # if os.path.exists(opt.save_dir):  # output dir
-    #     shutil.rmtree(opt.save_dir)  # delete dir
-    # os.makedirs(opt.save_dir)  # make new dir
     half = device.type != "cpu"  # half precision only supported on CUDA
 
     # Load model
@@ -63,10 +67,10 @@ def detect(cfg, opt):
     if opt.source.isnumeric():
         cudnn.benchmark = True  # set True to speed up constant image size inference
         dataset = LoadStreams(opt.source, img_size=opt.img_size)
-        bs = len(dataset)  # batch_size
+        _bs = len(dataset)  # batch_size
     else:
         dataset = LoadImages(opt.source, img_size=opt.img_size)
-        bs = 1  # batch_size
+        _bs = 1  # batch_size
 
     # Get names and colors
     names = model.module.names if hasattr(model, "module") else model.names
@@ -75,7 +79,7 @@ def detect(cfg, opt):
     # Run inference
     t0 = time.time()
 
-    vid_path, vid_writer = None, None
+    _vid_path, _vid_writer = None, None
     img = torch.zeros((1, 3, opt.img_size, opt.img_size), device=device)  # init img
     _ = model(img.half() if half else img) if device.type != "cpu" else None  # run once
     model.eval()
@@ -98,8 +102,7 @@ def detect(cfg, opt):
         t1 = time_synchronized()
         det_out, da_seg_out, ll_seg_out = model(img)
         t2 = time_synchronized()
-        # if i == 0:
-        #     print(det_out)
+
         inf_out, _ = det_out
         inf_time.update(t2 - t1, img.size(0))
 
@@ -124,7 +127,7 @@ def detect(cfg, opt):
         )
 
         _, _, height, width = img.shape
-        h, w, _ = img_det.shape
+        _h, _w, _ = img_det.shape
         pad_w, pad_h = shapes[1][1]
         pad_w = int(pad_w)
         pad_h = int(pad_h)
@@ -134,14 +137,12 @@ def detect(cfg, opt):
         da_seg_mask = torch.nn.functional.interpolate(da_predict, scale_factor=int(1 / ratio), mode="bilinear")
         _, da_seg_mask = torch.max(da_seg_mask, 1)
         da_seg_mask = da_seg_mask.int().squeeze().cpu().numpy()
-        # da_seg_mask = morphological_process(da_seg_mask, kernel_size=7)
 
         ll_predict = ll_seg_out[:, :, pad_h : (height - pad_h), pad_w : (width - pad_w)]
         ll_seg_mask = torch.nn.functional.interpolate(ll_predict, scale_factor=int(1 / ratio), mode="bilinear")
         _, ll_seg_mask = torch.max(ll_seg_mask, 1)
         ll_seg_mask = ll_seg_mask.int().squeeze().cpu().numpy()
         # Lane line post-processing
-        # ll_seg_mask = morphological_process(ll_seg_mask, kernel_size=7, func_type=cv2.MORPH_OPEN)
         # ll_seg_mask = connect_lane(ll_seg_mask)
 
         img_det = show_seg_result(img_det, (da_seg_mask, ll_seg_mask), _, _, is_demo=True)
@@ -161,24 +162,11 @@ def detect(cfg, opt):
         # if dataset.mode == 'images':
         cv2.imwrite(save_path, img_det)
 
-        p = str(opt.save_dir + "/" + Path(path).name)
+        name_lane = [Path(path).name, json.dumps(ll_seg_mask, cls=NumpyEncoder)]
+        name_lanes.append(name_lane)
 
-        n = p.split("/")[-1]
-
-        json_n = f"{n[0:n.rindex('.')]}.json"
-        json_o = os.path.join(opt.json_path, json_n)
-        print(f"Writing json to {json_o}")
-        with open(json_o, "w", encoding="utf-8") as jsonf:
-            json.dump(ll_seg_out.cpu().detach().numpy().tolist(), jsonf)
-
-        fields = ["img_name", "lanes_clean"]
-        csv_n = f"{n[0:n.rindex('.')]}.csv"
-        csv_o = os.path.join(opt.csv_path, csv_n)
-        print(f"Writing csv to {csv_o}")
-        with open(csv_o, "w") as f:
-            write = csv.writer(f)
-            write.writerow(fields)
-            write.writerows([ll_seg_out])
+    df = pd.DataFrame(name_lanes, columns=["source_image", "lanes"])
+    df.to_csv(path_or_buf=os.path.join(opt.csv_path, "lanes.csv"), index=False)
 
     print("Results saved to %s" % Path(opt.save_dir))
     print("Done. (%.3fs)" % (time.time() - t0))