+++ modules/sensor-extraction/ros-to-png/app.py
@@ -1,6 +1,9 @@
+# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# SPDX-License-Identifier: Apache-2.0
+
 import os
 
-from aws_cdk import App, CfnOutput, Environment
+from aws_cdk import App, CfnOutput, Environment, RemovalPolicy
 from stack import RosToPngBatchJob
 
 deployment_name = os.getenv("ADDF_DEPLOYMENT_NAME", "")
@@ -18,12 +21,13 @@ vcpus = int(os.getenv(_param("VCPUS"), 4))
 memory_limit_mib = int(os.getenv(_param("MEMORY_MIB"), 16384))
 resized_width = os.getenv(_param("RESIZED_WIDTH"))
 resized_height = os.getenv(_param("RESIZED_HEIGHT"))
+removal_policy = os.getenv(_param("REMOVAL_POLICY"), "")
 
 if resized_width:
-    resized_width = int(resized_width)
+    resized_width = int(resized_width)  # type: ignore
 
 if resized_height:
-    resized_height = int(resized_height)
+    resized_height = int(resized_height)  # type: ignore
 
 if not full_access_policy:
     raise ValueError("S3 Full Access Policy ARN is missing.")
@@ -44,8 +48,9 @@ stack = RosToPngBatchJob(
     vcpus=vcpus,
     memory_limit_mib=memory_limit_mib,
     s3_access_policy=full_access_policy,
-    resized_width=resized_width,
-    resized_height=resized_height,
+    removal_policy=RemovalPolicy.RETAIN if removal_policy.upper() == "RETAIN" else RemovalPolicy.DESTROY,
+    resized_width=resized_width,  # type: ignore
+    resized_height=resized_height,  # type: ignore
 )
 
 CfnOutput(